+++ packages/ironbean/src/core/decorators.ts
@@ -1,6 +1,6 @@
 import "reflect-metadata";
 import {
-    ApplicationContext,
+    ApplicationContext, ClassComponent,
     Component,
     ComponentType,
     constants,
@@ -19,7 +19,7 @@ export function component(ClassOrType: Class | ComponentType): any {
     let componentType = ComponentType.Singleton;
     function decorator(Class: any): any {
         Reflect.defineMetadata(constants.component, true, Class);
-        Reflect.defineMetadata(constants.componentType, componentType, Class);
+        Component.create(Class).setType(componentType);
 
         return Class;
     }
@@ -37,7 +37,7 @@ export function component(ClassOrType: Class | ComponentType): any {
 
 export function scope(scope: Scope): ClassDecorator {
     function decorator(Class: any): any {
-        Reflect.defineMetadata(constants.scope, scope, Class);
+        (Component.create(Class) as ClassComponent<any>).setScope(scope)
 
         return Class;
     }