+++ packages/ironbean/src/core/internals.ts
@@ -26,3 +26,4 @@ export * from "./utils";
 export * from "./types";
 export * from "./plugin";
 export * from "./storageMode";
+export * from "./inject";